Training 36-39
Set up agent to be rewarded depedent on whether the agent is facing the target and how far it is from the target, i.e., 1 - distanceFromTarget. This is to prevent agent from just gaining lots of reward by just looking at the target and encouraged movement. New best NN is 39